# api/openapi/releasepilot-v1.yaml
openapi: 3.1.0
info:
  title: ReleasePilot Control Plane API
  version: 0.1.0
  description: >-
    ReleasePilot MVP의 릴리스 요청, 승인, Canary 운영 명령,
    정책 및 감사 조회 API 계약이다.
servers:
  - url: /api/v1
    description: Same-origin API
...
paths:
...
  /session:
    get:
      tags: [Session]
      summary: 현재 세션 조회
      operationId: getCurrentSession
      responses:
        '200':
          description: 현재 사용자와 권한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /session/login:
    post:
      tags: [Session]
      summary: 사용자 로그인
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공. 응답 cookie로 서버 세션을 설정한다.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: 로그인 시도 제한 초과
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /session/demo:
    post:
      tags: [Session]
      summary: 읽기 전용 데모 세션 시작
      operationId: startDemoSession
      security: []
      responses:
        '200':
          description: VIEWER 데모 세션 생성됨
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '429':
          description: 데모 세션 발급 제한 초과
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /session/logout:
    post:
      tags: [Session]
      summary: 현재 세션 종료
      operationId: logout
      responses:
        '204':
          description: 로그아웃 완료
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /projects/{projectId}/memberships:
    post:
      tags: [Catalog]
      summary: 프로젝트 멤버 추가
      operationId: addProjectMembership
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddProjectMembershipRequest'}
      responses:
        '201':
          description: 멤버 추가됨
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectMembership'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
    get:
      tags: [Catalog]
      summary: 프로젝트 멤버 목록
      operationId: listProjectMemberships
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 프로젝트 멤버 목록
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ProjectMembership'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [Catalog]
      summary: 프로젝트 서비스 목록 조회
      operationId: listServices
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 서비스 목록
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                  nextCursor:
                    type: [string, 'null']
        '404':
          $ref: '#/components/responses/NotFound'
  /services/{serviceId}/environments:
    post:
      tags: [Catalog]
      summary: 서비스 환경 등록
      operationId: createEnvironment
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnvironmentRequest'
      responses:
        '201':
          description: DRAFT 환경 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /environments/{environmentId}/validate:
    post:
      tags: [Catalog]
      summary: 외부 리소스와 관측 계약 검증
      operationId: validateEnvironment
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: 검증 작업 접수됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationReceipt'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /environments/{environmentId}/validation-results/latest:
    get:
      tags: [Catalog]
      summary: 최근 환경 검증 결과 조회
      operationId: getLatestEnvironmentValidation
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '200':
          description: 최근 검증 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentValidation'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /policies:
    post:
      tags: [Policies]
      summary: 정책 생성
      operationId: createPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyRequest'
      responses:
        '201':
          description: 정책 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /policies/{policyId}/versions:
    post:
      tags: [Policies]
      summary: 정책 초안 버전 생성
      operationId: createPolicyVersion
      parameters:
        - name: policyId
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyVersionRequest'
      responses:
        '201':
          description: 정책 초안 버전 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersion'
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
  /policies/versions/{policyVersionId}/activate:
    post:
      tags: [Policies]
      summary: 정책 버전 활성화
      operationId: activatePolicyVersion
      parameters:
        - name: policyVersionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 활성화됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: RELEASEPILOT_SESSION
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EnvironmentId:
      name: environmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 같은 사용자와 operation에서 24시간 동안 중복 실행을 방지한다.
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: 요청 형식이 잘못됨
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 인증 필요
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: 권한 부족
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: 현재 상태 또는 동시 실행과 충돌
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: 형식은 유효하지만 의미 검증 실패
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          minLength: 1
          maxLength: 200
    Session:
      type: object
      required: [user, csrfToken, expiresAt]
      properties:
        user:
          $ref: '#/components/schemas/SessionUser'
        csrfToken:
          type: string
          description: unsafe method 호출 시 X-CSRF-TOKEN header로 보낸다.
        expiresAt:
          type: string
          format: date-time
    SessionUser:
      type: object
      required: [id, displayName, roles, demo]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        roles:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [VIEWER, DEVELOPER, APPROVER, OPERATOR]
        demo:
          type: boolean
...
    AddProjectMembershipRequest:
      type: object
      additionalProperties: false
      required: [userId, role]
      properties:
        userId: {type: string, format: uuid}
        role: {type: string, enum: [VIEWER, DEVELOPER, APPROVER]}
    ProjectMembership:
      type: object
      required: [id, projectId, userId, role, createdAt]
      properties:
        id: {type: string, format: uuid}
        projectId: {type: string, format: uuid}
        userId: {type: string, format: uuid}
        role: {type: string, enum: [VIEWER, DEVELOPER, APPROVER]}
        createdAt: {type: string, format: date-time}
...
    Service:
      type: object
      required: [id, projectId, key, name, repositoryUrl, owner, status]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        repositoryUrl:
          type: string
          format: uri
        owner:
          type: string
        status:
          type: string
          enum: [ACTIVE, DISABLED]
        createdAt:
          type: string
          format: date-time
...
    CreateEnvironmentRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - clusterId
        - namespace
        - rolloutName
        - containerName
        - stableServiceName
        - canaryServiceName
        - prometheusConnectionId
        - workloadLabelSelector
        - defaultPolicyVersionId
      properties:
        name:
          type: string
          enum: [staging, production]
        clusterId:
          type: string
          format: uuid
        namespace:
          type: string
          pattern: '^[a-z0-9]([-a-z0-9]*[a-z0-9])?$'
        rolloutName:
          type: string
        containerName:
          type: string
          description: Container in the Rollout pod template whose image is replaced
        stableServiceName:
          type: string
        canaryServiceName:
          type: string
        prometheusConnectionId:
          type: string
          format: uuid
        workloadLabelSelector:
          type: object
          minProperties: 1
          additionalProperties:
            type: string
        defaultPolicyVersionId:
          type: string
          format: uuid
    Environment:
      type: object
      required:
        - id
        - serviceId
        - name
        - clusterId
        - namespace
        - rolloutName
        - containerName
        - status
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        name:
          type: string
          enum: [staging, production]
        clusterId:
          type: string
          format: uuid
        namespace:
          type: string
        rolloutName:
          type: string
        containerName:
          type: string
        status:
          type: string
          enum: [DRAFT, VALIDATING, ACTIVE, ACTIVE_WITH_WARNINGS, INVALID, DISABLED]
        lastValidatedAt:
          type: [string, 'null']
          format: date-time
    ValidationReceipt:
      type: object
      required: [validationId, environmentId, status, acceptedAt]
      properties:
        validationId:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        status:
          const: ACCEPTED
        acceptedAt:
          type: string
          format: date-time
    EnvironmentValidation:
      type: object
      required: [id, environmentId, status, checks, startedAt]
      properties:
        id:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        status:
          type: string
          enum: [RUNNING, PASSED, PASSED_WITH_WARNINGS, FAILED]
        checks:
          type: array
          items:
            $ref: '#/components/schemas/ValidationCheck'
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: [string, 'null']
          format: date-time
    ValidationCheck:
      type: object
      required: [key, result, message]
      properties:
        key:
          type: string
        result:
          type: string
          enum: [PASS, WARNING, FAIL]
        message:
          type: string
        details:
          type: object
...
    CreatePolicyRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    Policy:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        status: {type: string, enum: [ACTIVE, DISABLED]}
        createdAt: {type: string, format: date-time}
    CreatePolicyVersionRequest:
      type: object
      additionalProperties: false
      required: [definition]
      properties:
        definition:
          type: object
          description: release-policy.schema.json을 따라야 한다.
    PolicyVersion:
      type: object
      required: [id, policyId, version, status, checksum, createdAt]
      properties:
        id:
          type: string
          format: uuid
        policyId:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [DRAFT, ACTIVE, RETIRED]
        checksum:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Problem:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        traceId:
          type: string
        violations:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
security:
  - sessionCookie: []